Avoid messing with 'doc' images
...when running specific tests with 'dev-env/end-to-end-tests.sh --spec'

+++ dev-env/end-to-end-tests.sh
@@ -31,7 +31,10 @@ docker compose exec \
     npx cypress run \
       --e2e "$@"
 
-find ../doc/user -name '*.png' -delete
-find ../tests/screenshots/gabby.cy.js -name '*.png' | while read line; do
-  cp $line ../doc/user/${line#../tests/screenshots/gabby.cy.js/}
-done
+if [ -d ../tests/screenshots/gabby.cy.js ]
+then
+  find ../doc/user -name '*.png' -delete
+  find ../tests/screenshots/gabby.cy.js -name '*.png' | while read line; do
+    cp $line ../doc/user/${line#../tests/screenshots/gabby.cy.js/}
+  done
+fi
